Migrate pip.sh into piptool.py
This migrates the logic from pip.sh into piptool.py, which should improve portability by removing the bash dependency.

This also has the beginnings of wrapping piptool as a closed redistributable that doesn't rely on a system-installed copy of PIP, but instead uses these rules to pull pip into a PAR bundle.  Besides needing to work out the details of releasing and redistributing the PAR, we have two unresolved issues:
 * When bundled as a PAR (vs. py_binary), piptool seems to pick up the system-installed version of pip.
 * When bundled as a PAR, piptool sometimes sees cert issues resolving requirements (similar to what we see with httplib2).