Upgrade documentation generation to Stardoc
This was a long time coming. Our previous Skydoc workflow no longer worked, and
Skydoc has been deprecated for a while.

This commit updates the WORKSPACE deps for Stardoc, and rewrites the docgen
workflow in docs/BUILD and update_docs.sh. Note that running the workflow
requires some visibility changes to @bazel_tools, which will be in Bazel 0.29.

Fixes #144.