Add a nice error message to help with the workspace name migration
This hooks into pip_repositories(), which users are *supposed* to be calling in
their WORKSPACE files, to emit a nice fail() message alerting them that they
need to update their repo definition.

Without this change (and even with it, for users who do not call
`pip_repositories()`), users will instead see a confusing cyclic dependency
error.