Migrate to stardoc and regenerate documentation (#213)
* Upgrade documentation generation to Stardoc

This was a long time coming. Our previous Skydoc workflow no longer worked, and
Skydoc has been deprecated for a while.

This commit updates the WORKSPACE deps for Stardoc, and rewrites the docgen
workflow in docs/BUILD and update_docs.sh. Note that running the workflow
requires some visibility changes to @bazel_tools, which will be in Bazel 0.29.

Fixes #144.

* Regenerate documentation using Stardoc

This rips out the old documentation, which was malformed and unreproducible,
and adds documentation generated by Stardoc.

It was generated using a development version of Bazel that makes some
@bazel_tools targets publically visible. These changes will be in Bazel 0.29.

Follow-up for #144.

* Exclude docs/ from CI config

The new docgen workflow requires Bazel 0.29, which is currently unreleased.
This commit excludes the docgen from being tested on CI. We'll undo this
after 0.29 is released (#144).