Add pip3_import
This adds two repo rule macros, `pip2_import` and `pip3_import`, as wrappers
around `pip_import` that set `python_interpreter` to "python2" and "python3"
respectively. The latter is important for requesting the Python 3 version of a
package, since the `pip_import` rule invokes "python", which is Python 2 on
most systems.

piptool.py is updated to prefix the names of the wheel repos (an implementation
detail of rules_python) with the name given to `pip_import`. This is needed to
avoid shadowing wheel repos when the same wheel name is used by separate
`pip_import` invocations -- in particular when the same wheel is used for both
PY2 and PY3. (Thanks to @joshclimacell for pointing this detail out in
his prototype 90a70d5a550a01646966ce61156e6f83f02b4b73.)

Regenerated the .par files and docs.