Add python_interpreter attr to pip rules (#252)
This adds a `python_interpreter` attribute to `pip_import` and `whl_library` that can be used to select the system command used to run Python's packaging tools. This provides the basis for invoking pip under Python 3 to install PY3 dependencies.

Example usage:
```
pip_import(
    name = 'pip_deps',
    requirements = '//:requirements.txt',
    python_interpreter = 'python3.7',
)
```

The par files have been regenerated. (This required a little bootstrapping since piptool.py needs to be modified to accept the flag before pip.bzl is modified to pass it.)