chore: github setup improvements
See https://github.com/bazelbuild/rules_python/community for the items we are missing.

Issues and PRs:
See https://docs.github.com/en/github/building-a-strong-community/using-templates-to-encourage-useful-issues-and-pull-requests
This will ask users filing an issue to indicate if it's a bug or a feature request and prompt them to include answers to common questions, reducing how many round-trips they need from maintainers.
Same for pull requests, especially to think about whether this is a breaking change.

Also add a .bazelversion file to pin the version devs and CI use when developing in the repo.

In the future we might want rules_codeowners to generate the .github/CODEOWNERS file as we expand community maintainers