ci: add bazel_integration_test
This runs a py_test with a copy of bazel as a data dep.
It glob()s up the sources for each example and runs nested bazel test on them.
This detects whether the examples are fully working and self-contained.

Follow-up step is to replace the rules_python.tgz with a HEAD version so we detect
breakages.