Fix extract_single_wheel for Windows (#498)
On Windows, when using pip_parse, pip would fail with following error:

    Could not open requirements file: [Errno 13] Permission denied: ...

This is due to Python holding a handle to the temporary file preventing
pip, which is run as a sub-process, from reading it. For more
information, see: https://bugs.python.org/issue14243

Closing the requirements file before running pip solves the problem.