Don't include pyc files as part of the hermetic toolchain
pyc files don't appear to be generated deterministically, so including
them as part of the :files filegroup means that actions that use the
hermetic toolchain do not have deterministic hashes, and so bazel
needlessly re-executes them. This can be seen when using bazel together
with a remote cache: builds on separate machines will generate pyc files
with different hashes, leading to cache misses.