Normalize distribution name with pkg_resources
`pypa/installer` is used to parse Wheel metadata, but does not currently
provide a method for normalizing distribution names:

- https://github.com/pypa/installer/issues/97

`pypa/pkg_resources` provides `Requirement.parse` which returns an instance
of `Requirement` where `.key` is the canonical distribution name per PEP 503.

The `Requirement` class can also parse `extras`, but it returns a normalized
form that I believe could break the installation of the extras.