feat: add alias generation for pip parse
Generate alias packages and targets for `pip_parse` under the parent repository.
This serves two purposes. It allows for users to reference packages using a
simpler `@pypi//package_name` without the import issues associated with
putting the contents of the package there (as may be the case with `pip_install`
following this naming pattern). Additionally, this allows users to override
individual libraries either by default or based on select statements.