Support annotations on pip packages with extras.
E.g., the following requirement:

```
requests[security]>=2.8.1
```

This is handled correctly by all of the other plumbing, but trying
to add an annotation to `requests` will fail. This is because
annotations have separate logic for parsing requirements in the
generated .bzl file. It would previously turn the requirement into
`requests[security]` rather than just `requests`.