runfiles: Apply repo mapping to Rlocation path
When a repo mapping manifest is available in runfiles, it is parsed and
used to map apparent repository names to canonical ones in paths passed
to Rlocation. The current repository, which is required to know which
part of the mapping to apply, is either determined using
`CurrentRepository` or can be passed in explicitly.

With this commit, runfiles lookups should succeed with Bzlmod without
code changes.