refactor(example/bzlmod): mv __init__.py lib.py
The `__init__.py` files in the root of the WORKSPACE in `bzlmod` is
breaking, when running under `bazel coverage //:test`. However, it
started working when I renamed `__init__.py` to `lib.py`. I am
suspecting that this has to do with the fact that the layer of
indirection that `coverage` introduces could be something to do with
that.