feat(coverage): Set visibility to the toolchain repo only for non-bzlmod
This changes how we install the coverage dependencies and allows us to
set the visibility for the coverage tool correctly by passing the name
of the toolchain repo during the repository function execution.

bzlmod does not have this due to the fact that the dependency
registering is happening in a different way/order.