feat!: generate user friendly aliases for everyone
This is related to #814 as it is bringing user-friendly alias generation
and it is doing this in a breaking way for the `bzlmod` users who are
not using the `requirement` macro. The labels that they use will change
from `@@repo//:{name}_pkg` to `@@repo//{name}`.

I am intentionally leaving the requirement macros in the
requirements.bzl unchanged as #989 might change the approach again.

Since these aliases can be used by bzlmod and non-bzlmod users, we can
then support the same generated `gazelle_python.yaml` for both bzlmod
and non-bzlmod users.