fix: move rules_proto dependencies to py_repositories
Currently the `py_proto_library` depends on protobuf toolchain and this
is present only in the `internal_deps`. Whilst trying to use
`py_proto_library` we noticed that we need to add the dependencies
ourselves.

This change moves the location of the dependency declaration to the
function that is gonna be called by our users.