Make toolchain acceptance tests work with latest Bazel build CI pipeline.
The latest Bazel build continous integration testing pipeline sets
several flags and environment variables that end up interfering with
each other:
 * `--sandbox_tmpfs_path=/tmp`
 * `--test_env=USE_BAZEL_VERSION`
 * `USE_BAZEL_VERSION=/tmp/<something>`
 * And Bazelisk is used to run Bazel

What happens is `USE_BAZEL_VERSION` points to Bazel in /tmp, but then
the `--sandbox_tmpfs_path` flag prevents it from being readable. Later,
when a test wants to run Bazel, Bazelisk is invoked. It is able to see
that it should use a custom Bazel binary because of `--test_env`, but
then can't read the file because of `--sandbox_tmpfs_path`, so then
fails.

To fix, make the test runner that will run `bazel` unset
`USE_BAZEL_VERSION` so Bazelisk doesn't try to use it.

This also exposed an issue with Bazelisk demanding a cache directory be
specified, so set that environment variable to the test's temp dir to
keep Bazelisk happy.

Fixes #856