build(precommits): Add commitizen to pre-commit hooks to enforce commit message formatting
Having a standard format for commit messages makes it easier to
understand what general behaviors and functionality are being affected.
This makes it easier to easier to generate changelog information for
releases and quickly skim PRs.

The default commitizen style is used, which is Conventional
Commits. Docs are updated to give a brief description of the format
and common types.