Updating example to only use bzlmod
This commit updates the gazelle example to only use bzlmod.  The
example was previously set up to support bzlmod and also
pre-bzlmod configurations, but moving forward I recommend just
using bzlmod.  We also missed a change in gazelle because
bzlmod was picking up elements out of the WORKSPACE file.

This commit removes the CI for the legacy bazel support as well.