tests: Set linux platform for test_non_mac_doesnt_require_darwin_for_execution
This is the inverse test of the one testing for the mac platform. As
before, the line forcing the platform to a non-mac platform was
commented out, which meant, when a mac host built it, it would build the
underlying target for Mac, which violated the test's assumptions.

Work towards bazelbuild/bazel/issues/18170