tests: Force analysis test labels to resolve within @rules_python context
When a string label is passed to the `@rules_testing` analysis_test
functions, the strings are evaluated within the context of
@rules_testing because that is where the actual rule invocation happens.
Without bzlmod, this just requires qualifying the labels with the repo
name (which is what was being done) because there's just a flat global
namespace of repos.

With bzlmod enabled, repo mapping happens, so rules_testing tries to
resolve those repo names using its repo mapping, which doesn't work
because rules_testing's mapping doesn't include every repo using it.

To fix, we can force repo resolution to happen in rule_python's context
by calling Label().