fix(update_deleted_packages.sh): allow to run from anywhere in the repo (#1206)
It seems that the tool was correctly trying to modify the `.bazelrc` at
the
root of the repo, but the `find` execution would run in the `$PWD`. This
change ensures that the `find` is consistent with the file we are trying
to modify and allows the user to execute the script from anywhere in the
repo.

Tested:
1. Update the deleted packages with the version of the script at HEAD
1. Fix the bug
1. Rerun the script from the 'tests' sub-folder in the repo to ensure
that
   running the script is noop.
1. Revert the changes to '.bazelrc' to reduce conflicts as other PRs are
   modifying it.

Work towards #958.