feat(bzlmod): Cleaning up interpreter resolution
This commit cleans up the use of "canonical resolution" of
the Python interpreter. When the extension toolchains run
it collects a list of the interpreters and then
uses the hub_repo rule to create a map of names and the
interpreter labels.

Next, we then use the interpreter_extension that, creates
reports that have symlinks pointing to the different interpreter
binaries.

The user can then pass in a label to the pip call for the
specific hermetic interpreter.