feat(bzlmod): a minimal entry_point support and a test
Add an `entry_point` macro to the repo generated by the `pip.parse` extension.
This works by using the canonical label literal, so should work without users
needing to add any "spoke" repos to the `use_repo` statement. We test this by
having an extra folder in the `bzlmod` example.