test(bzlmod): explicitly enable bzlmod in the test harness (#1204)
Previously we would depend on the value of .bazelrc and this
change ensures that we are explicitly enable bzlmod via CLI
args. It seems that the `py_proto_library` integration tests
defined in the `//examples:BUILD.bazel` file were not running
using `bzlmod` before hand, however, they were correctly
executed in the CI.

Work towards #958.