feat(gazelle): allow passing multiple files to the manifest generator
For certain workflows it is useful to calculate the integrity hash of
the manifest file based on a number of requirements files. The
requirements locking is usually done by executing a script on each
platform and having gazelle manifest generator be aware that more than
one requirements file may affect the outcome (e.g. the wheels that get
passed to modules map may come from `multi_pip_parse` rule) is generally
useful.

This change modifies the `generator` to support multiple
`--requirements` flag specification and uses all files that got passed
to generate the final integrity hash.