feat(bzlmod): Allowing multiple python.toolchain extension calls (#1230)
We do this work for two reasons.

First, we must support Module dependencies and sub-modules using
`python.toolchain`. There are already two known instances of sub-modules
setting up a Python toolchain and colliding with another module
(nanobind and rules_testing both run into this).

Second, the upcoming multi-version support is going to work by having
each `python.toolchain()` call register its particular version with the
extra toolchain constraint. This also helps unify the version-aware and
non-version-aware code paths (the non-version aware paths are just
version-aware with a single version registered as the default)

This commit implements various business logic in the toolchain class.

Toolchains in Sub Modules

It will create a toolchain in a sub-module if the toolchain
of the same name does not exist in the root module. The extension stops
name
clashing between toolchains in the root module and sub-modules.
You cannot configure more than one toolchain as the default toolchain.

Toolchain set as the default version.

This extension will not create a toolchain in a sub-module
if the sub-module toolchain is marked as the default version. If you
have
more than one toolchain in your root module, you need to set one of the
toolchains as the default version.  If there is only one toolchain, it
is set as the default toolchain.

See #1229 for more information