feat(bzlmod): Moving register.toolchains internal (#1238)
This commit moves the register.toolchains bzlmod call to inside
of rules_python.  Instead of a user having to call register.toolchains
in their MODULE.bazel, rules_python/MODULE.bazel calls it
on the internal hub.

This is a breaking change if you are using register.toolchains inside
of submodules.  Using register.toolchains inside of submodules is
not recommended anyways.  This is now broken because we are not
creating a repo for every Python version toolchain.  All of the
toochain calls exist now in the hub's repo BUILD.bazel file.