cleanup: Set toolchain target_setting directly instead of via inline ternary
The generated toolchain BUILD file is confusing to read because it
relies on a ternary expression in the BUILD file to set the
`target_settings` attribute. This makes debugging harder because,
upon first look, all the toolchains appear to have the version
constraint set. It's only upon closer inspection that you can see the
1-character difference of "False" vs "True" embedded into the middle
of a line amongst other similar looking lines.

Also adds a bit of validation logic for the
`set_python_version_constraint` argument because it's conceptually a
boolean, but is passed as a string, so is prone to having an incorrect
value passed.