feat(pip): Code simplification
Working on doing some cleanup of the pip_repository
and pip extension code. This refactor merges the two
seperate code paths we had for bzlmod vs non-bzlmod
implementation.

This commit also includes refatoring so that we do not
have multiple for loops over the same lists and other
code cleanup.