fix(bzlmod): give precedence to the first seen versioned toolchain (#1244)
This fixes an issue where the last submodule, instead of the first, was
given precedence when creating versioned toolchains. To fix, when
creating the map of versioned toolchains, if a version is already
defined, then a subsequent usage is ignored. A warning is emitted when
this occurs.

This also fixes a similar problem that can occur to the root module. If
the root module uses a particular version marked as the default, and is
using the versioned rules, and a submodule also uses that same version,
then the submodule's toolchain would be used. This happened because the
root module's toolchain would be moved last, so its versioned rules
would match the submodule's versioned toolchain.

This also does some cleanup and refactoring to:
  * Compute the toolchains in one loop iteration
  * Give more informative error messages
* Reject duplicate names within a module, even for the non-root module.
  * Reject duplicate versions within a module.