fix(coverage): patch sys.path before importing anything from coverage
The previous patch was changing the `sys.path` in the
`coverage/cmdline.py` and it seems that it was impossible to get
coverage working because of a circular import error.  This was because
Python would attempt to load from the `coverage/types.py` when instead
of `types` from the stdlib because of the ordering of `sys.path`.

This patch now works around the issue by sorting it before anything
other than `__main__.py` gets processed by the Python import machinery.