refactor(coverage): register coverage dependencies next to the toolchain
This unifies how the coverage toolchain is registered under bzlmod and
non-bzlmod by leveraging the fact that repositories created by the same
bzlmod extension can see each other. This means that the `MODULE.bazel`
file can be cleaner and the visibility of the `coverage` tool can be
restricted to be only visible by the appropriate toolchain.

This also means that we have to relax the 'coverage_tool' attribute for
the python_repository to be a simple string as the label would be
evaluated within the context of `rules_python` which may not necessarily
have access to it via the friendly name used from the toolchain repo.