feat: Adding variable support for distribution in py_wheel (#1251)
This allows the `distribution` attribute to expand workspace status
keys, just as the
`version` attribute can.

This allows, for example, the VCS's branch name (e.g test, release, etc)
to be part of the
distribution name without having to modify the BUILD file. Having
distinct distribution
names is necessary because tools like pip-compile, which determine
version compatibility
and replacements, and having the same distribution name would mean the
different builds
are seen as equivalent.

Closes #1250