Fix(bzlmod): Windows Python Interpreter symlink issues
When using Windows you cannot run the symlink directly.

Because of how symlinks work in Windows we need to use the path
realink.
Unlike Linux and OSX we cannot execute the Python symlink directly.
Windows throws an error "-1073741515", when we execute the symlink
directory.  This error means that the Python binary cannot find dlls.
This is because the symlink that we create is not in the same folder
as the dlls.

This commit introduces code that resolves the actual location of the
interpreter using the path realink when we are running bzlmod and
Windows.