feat(bzlmod): Implementing wheel annotations/whl_mods
This commit implements a bzlmod extension that allows users
to create "annotations" for wheel builds.  The wheel_builder.py
accepts a JSON file via a parameter called annotations, and
this extension creates those JSON files. The pip extension accepts
a Label -> String dict argument of said JSON files.

We renamed this to whl_mods because we are handling the JSON files
a bit differently, and IMO annotations is a bit confusing.  We
are providing modifications to the creation of the wheel build
file and other components that impact the wheel build.